+++ js/jquery.terminal-src.js
@@ -246,9 +246,10 @@
     function DelayQueue() {
         var callbacks = $.Callbacks();
         var resolved = false;
+        var self = this;
         this.resolve = function() {
             callbacks.fire();
-            resolved = true;
+            self.resolved = resolved = true;
         };
         this.add = function(fn) {
             if (resolved) {
@@ -3050,6 +3051,7 @@
             } else {
                 prompt = prompt_node.text();
             }
+            prompt = prompt.replace('\u200b', '');
             prompt = $.terminal.escape_brackets(prompt);
             var prompt_marker = get_prompt_marker();
             var re = new RegExp('^' + prompt_marker + $.terminal.escape_regex(prompt));
@@ -3510,7 +3512,7 @@
                 var lines = $.terminal.split_equal(tmp_prompt, num_chars);
                 lines = lines.map(function(line) {
                     return line.replace(/^\uFFFF+/, '');
-                }).filter(Boolean);
+                });
                 lines = lines.map(function(line) {
                     if (!$.terminal.have_formatting(line)) {
                         return '[[;;]' + $.terminal.escape_brackets(line) + ']';
@@ -3541,10 +3543,12 @@
                            '</span>';
                 }).concat([last_line]).join('\n');
             }
-            function set(prompt) {
+            function set(prompt, options) {
                 if (prompt) {
-                    prompt = $.terminal.apply_formatters(prompt, {prompt: true});
-                    prompt = $.terminal.normalize(prompt);
+                    if (options && options.formatters || !options) {
+                        prompt = $.terminal.apply_formatters(prompt, {prompt: true});
+                        prompt = $.terminal.normalize(prompt);
+                    }
                     prompt = crlf(prompt);
                 }
                 var formatted = format_prompt(prompt);
@@ -3574,7 +3578,7 @@
                     }
                 }
             }
-            return function() {
+            return function(options) {
                 // the data is used as cancelable reference because we have ref
                 // data object that is hold in closure and we remove `set` function
                 // so previous call to function prompt will be ignored
@@ -3587,7 +3591,7 @@
                     set: set
                 };
                 with_prompt(prompt, function(prompt) {
-                    data.set(prompt);
+                    data.set(prompt, options);
                 }, self);
             };
         })();
@@ -3800,7 +3804,7 @@
                 prompt_offset = len;
                 prompt_len = just_prompt_len + prompt_offset;
             },
-            prompt: function(user_prompt) {
+            prompt: function(user_prompt, options) {
                 if (user_prompt === true) {
                     return last_rendered_prompt;
                 } else if (user_prompt === undefined) {
@@ -3814,7 +3818,7 @@
                         throw new Error('prompt must be a function or string');
                     }
                     if (should_redraw) {
-                        draw_prompt();
+                        draw_prompt(options);
                         // we could check if command is longer then numchars-new
                         // prompt
                         redraw();
@@ -7103,6 +7107,20 @@
             'params': options.params,
             'id': ++ids[options.url]
         };
+        function finalize(json, response, status, jqXHR) {
+            if (validJSONRPC(json) || options.method === 'system.describe') {
+                // don't catch errors in success callback
+                if (options.success) {
+                    options.success(json, status, jqXHR);
+                }
+                deferred.resolve(json);
+            } else {
+                if (options.error) {
+                    options.error(jqXHR, 'Invalid JSON-RPC');
+                }
+                deferred.reject({message: 'Invalid JSON-RPC', response: response});
+            }
+        }
         $.ajax({
             url: options.url,
             beforeSend: function beforeSend(jxhr, settings) {
@@ -7132,18 +7150,16 @@
                 if (is_function(options.response)) {
                     options.response(jqXHR, json);
                 }
-                if (validJSONRPC(json) || options.method === 'system.describe') {
-                    // don't catch errors in success callback
-                    if (options.success) {
-                        options.success(json, status, jqXHR);
+                if (is_function(options.intercept)) {
+                    var ret = options.intercept(request, json);
+                    if (ret) {
+                        return unpromise(ret, function(json) {
+                            var response = JSON.stringify(json);
+                            finalize(json, response, status, jqXHR);
+                        });
                     }
-                    deferred.resolve(json);
-                } else {
-                    if (options.error) {
-                        options.error(jqXHR, 'Invalid JSON-RPC');
-                    }
-                    deferred.reject({message: 'Invalid JSON-RPC', response: response});
                 }
+                finalize(json, response, status, jqXHR);
             },
             error: options.error,
             contentType: 'application/json',
@@ -7366,6 +7382,8 @@
     // -----------------------------------------------------------------------
     // :: Default options
     // -----------------------------------------------------------------------
+    // if set to false nested formatting will not process formatting, only text
+    // between formatting, we need this option because we're flattening the formatting
     $.terminal.nested_formatting.__meta__ = true;
     // if set to false nested formatting will not inherit styles colors and attribues
     $.terminal.nested_formatting.__inherit__ = true;
@@ -7704,6 +7722,12 @@
         // :: Create interpreter function from url string
         // ---------------------------------------------------------------------
         function make_basic_json_rpc(url, auth) {
+            var rpc_interceptor;
+            if (settings.rpc) {
+                rpc_interceptor = function() {
+                    return settings.rpc.apply(self, arguments);
+                };
+            }
             var interpreter = function(method, params) {
                 self.pause(settings.softPause);
                 $.jrpc({
@@ -7717,6 +7741,7 @@
                             display_exception(e, 'USER');
                         }
                     },
+                    intercept: rpc_interceptor,
                     response: function(jxhr, response) {
                         try {
                             settings.response.call(self, jxhr, response, self);
@@ -7996,7 +8021,6 @@
                         var rest = interpreters.slice(1);
                         var type = get_type(first);
                         if (type === 'string') {
-                            self.pause(settings.softPause);
                             if (settings.describe === false) {
                                 if (++rpc_count === 1) {
                                     fn_interpreter = make_basic_json_rpc(first, login);
@@ -8005,6 +8029,7 @@
                                 }
                                 recur(rest, success);
                             } else {
+                                self.pause(settings.softPause);
                                 make_json_rpc_object(first, login, function(new_obj) {
                                     if (new_obj) {
                                         $.extend(object, new_obj);
@@ -9099,13 +9124,14 @@
         // ---------------------------------------------------------------------
         function typed(finish_typing_fn) {
             return function typing_animation(message, options) {
-                var formattted = $.terminal.apply_formatters(message, {
+                var formatted = $.terminal.apply_formatters(message, {
                     animation: true
                 });
+                formatted = $.terminal.normalize(formatted);
                 animating = true;
                 var prompt = self.get_prompt();
                 var char_i = 0;
-                var len = $.terminal.length(formattted);
+                var len = $.terminal.length(formatted);
                 if (message.length > 0) {
                     var new_prompt = '';
                     if (options.prompt) {
@@ -9114,7 +9140,7 @@
                         self.set_prompt('');
                     }
                     var bottom = self.is_bottom();
-                    var chars = $.terminal.partition(formattted, {wrap: false});
+                    var chars = $.terminal.partition(formatted, {wrap: false});
                     var interval = setInterval(function() {
                         if (!skip) {
                             var chr = chars[char_i];
@@ -9127,16 +9153,16 @@
                                 }
                             }
                             new_prompt += chr;
-                            self.set_prompt(new_prompt);
+                            command_line.prompt(new_prompt, {formatters: false});
                             if (bottom && (chr === '\n' || !self.is_bottom())) {
                                 self.scroll_to_bottom();
                             }
                             char_i++;
                         } else {
                             self.skip_stop();
-                            var chrRest = $.terminal.substring(formattted, char_i, len);
-                            new_prompt += chrRest;
-                            self.set_prompt(new_prompt);
+                            var chr_rest = $.terminal.substring(formatted, char_i, len);
+                            new_prompt += chr_rest;
+                            command_line.prompt(new_prompt, {formatters: false});
                             char_i = len;
                         }
                         if (char_i === len) {
@@ -10449,6 +10475,9 @@
                                 wrapper.toggleClass('partial', appending_to_partial);
                                 if (appending_to_partial) {
                                     partial = wrapper;
+                                } else if (data.newline && partial.length) {
+                                    wrapper = $('<div/>');
+                                    partial = $();
                                 }
                                 data.finalize(wrapper);
                             } else {
@@ -11346,7 +11375,7 @@
                     }
                     $(window).off('blur', blur_terminal).
                         off('focus', focus_terminal);
-                    self.find('.terminal-fill, .terminal-font')
+                    self.find('.terminal-fill, .terminal-font, .terminal-font-forcer')
                         .remove();
                     self.stopTime();
                     terminals.remove(terminal_id);
@@ -11595,6 +11624,7 @@
             requests.push(xhr);
         });
         var scroller = $('<div class="terminal-scroller"/>').appendTo(self);
+        $('<div class="terminal-font-forcer terminal-hidden">x<div>').appendTo(self);
         var wrapper = $('<div class="terminal-wrapper"/>').appendTo(scroller);
         $(broken_image).hide().appendTo(wrapper);
         var font_resizer = $('<div class="terminal-font">&nbsp;</div>').appendTo(self);
@@ -12152,7 +12182,7 @@
                 }, options);
             }
             function bottom_detect(intersections) {
-                is_bottom_detected = intersections[0].intersectionRatio === 1;
+                is_bottom_detected = intersections[0].intersectionRatio >= 0.9;
             }
             function create_bottom_detect() {
                 if (window.IntersectionObserver) {
@@ -12245,7 +12275,12 @@
                 // don't make sense
                 observe_visibility();
             }
-            command_queue.resolve();
+            // wait for custom font to load #892
+            if (document.fonts && document.fonts.ready) {
+                document.fonts.ready.then(command_queue.resolve);
+            } else {
+                command_queue.resolve();
+            }
             // touch devices need touch event to get virtual keyboard
             if (enabled && self.is(':visible') && !is_mobile) {
                 self.focus(undefined, true);