+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.37.0
+### Features
+* add `rpc` interceptor [#883](https://github.com/jcubic/jquery.terminal/issues/883)
+* wait for the custom fonts to load [#892](https://github.com/jcubic/jquery.terminal/issues/892)
+### Bugfix
+* fix CMD wrapping when prompt is empty [#882](https://github.com/jcubic/jquery.terminal/issues/882)
+* fix paused terminal when using array as interpreter and RPC without system describe
+* fix edge case of disappearing content after refresh [#871](https://github.com/jcubic/jquery.terminal/issues/871)
+* fix rendering empty lines prompt and during typing animation [#890](https://github.com/jcubic/jquery.terminal/issues/890)
+* fix rendering links in prompt and during typing animation [#891](https://github.com/jcubic/jquery.terminal/issues/891)
+
 ## 2.36.0
 ### Features
 * split_equal accept optional object as 3rd argument with two options `trim` and `keepWords`